% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawards.R
\docType{package}
\name{metawards}
\alias{metawards}
\title{metawards: Metapopulation Disease Modelling Package}
\description{
This package provides a convenient R wrapper around the Python
metawards package. For more information about MetaWards see
\url{https://metawards.org}. For a quick-start guide see
\url{https://metawards.org/quickstart}.
}
\details{
This package uses reticulate to automatically load and wrap
the Python metawards package. To install metawards you will need
to either manually install metawards into the python interpreter
used by reticulate, or you can use the convenience function
'py_install_metawards()', which will install into the default
python environment supplied with reticulate.
}
\section{Python functions}{

The python functions that are wrapped are all documented
at \url{https://metawards.org/api}.
}

\section{R-specific functions}{

\itemize{
\item \code{\link{py_install_metawards}}
\item \code{\link{py_metawards_available}}
\item \code{\link{py_version_metawards}}
}
}

