% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawards.R
\name{py_metawards_available}
\alias{py_metawards_available}
\title{Return whether MetaWards is installed and available}
\usage{
py_metawards_available()
}
\value{
Boolean of whether MetaWards is installed
}
\description{
Return whether MetaWards is installed and available
}
