% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawards.R
\name{py_version_metawards}
\alias{py_version_metawards}
\title{Return the version number of the installed MetaWards module}
\usage{
py_version_metawards()
}
\value{
The version number as a string
}
\description{
Return the version number of the installed MetaWards module
}
