\name{install_metawards}
\title{Install MetaWards}
\usage{
install_metawards(method = "auto", python_version = 3.7)
}
\description{
Installs MetaWards into the Python environment associated with this R
session. Note that MetaWards requires Python 3.7 or above and will
be installed using pip.
}
\examples{
install_metawards()
}
